+++ tests/Stubs/Config.php
@@ -1,0 +1,14 @@
+<?php
+
+declare(strict_types=1);
+
+namespace ComposerUnused\SymbolParser\Test\Stubs;
+
+final class Config
+{
+    public string $name = '';
+    /** @var array<string, mixed> */
+    public array $autoload = [];
+    /** @var array<string, mixed> */
+    public array $require = [];
+}